Publish queue report (#6)
* Initial setup of Publish Queue reporting.

* Added changelog. Updated README. Added POC code for Publish Queue report.

* Fixed missing Font Awesome reference.

* Added Unicorn to project. Renamed "SPRK Publish Logs" to "SPRK Publish Reports".

* Created Index page for reports list. Moved old index page to PublishLogs page.

* Re-formatted report views to support multi-report concept.

* Structural changes to Main Menu.

* Report index page styling.

* Report index updates.

* Added web database configuration item. Style updates.

* ReportService. Created workflow-sensitive implementation of publish queue report.

* Beta version logic for Publish Queue report.

* Publish Queue report styling and language version updates.

* Added null check to Publish Queue report on scItem in case of deletion before publishing on source DB.

* Fixed missing language version bug.

* Styling tweaks to report index page.

* Added buffer flushing to Spark Logger.

* Added report index footer.

* Added Dispose method to ReportController.

* Switched ILogManager service to Transient lifetime instead of scoped.

* Set item ID column on ViewLog to 40% width.

* README updates.

* Added "Results" count to several reports.

* Updated AssemblyInfo to 1.5.0.

* Restructured stylesheets and fonts.

* Updated JS file locaiton.

* Added Sitecore package definition XML.

* Removed README link from readme section of Sitecore package XML.

* Fixed several issues with logger unit tests.

* Added missing processor files.

+++ SitecoreSpark.SPRK.Tests/SparkLoggerTests.cs
@@ -10,7 +10,9 @@ namespace SitecoreSpark.SPRK.Tests
     public class SparkLoggerTests
     {
         internal string _logFolderPath = "./TESTLOGS";
-        internal string _logFilePrefix = "SPRK.TESTLOG.";
+
+        internal string _logFilePrefix = "SPRK.TESTLOG";
+      
         internal int _logBufferSize = 1000;
 
         private SparkLogger SetupLogger()
@@ -21,11 +23,14 @@ namespace SitecoreSpark.SPRK.Tests
 
         private string GetTestFilePath()
         {
-            return Path.Combine(_logFolderPath, ($"{this._logFilePrefix}{DateTime.Now.ToString("yyyyMMdd")}.txt"));
+            return Path.Combine(_logFolderPath, ($"{this._logFilePrefix}.{DateTime.Now.ToString("yyyyMMdd")}.txt"));
         }
 
         private void CleanUpTestLogFiles()
         {
+            if (!Directory.Exists(this._logFolderPath))
+                return;
+
             foreach (var filePath in Directory.GetFiles(this._logFolderPath))
             {
                 File.Delete(filePath);
@@ -76,6 +81,8 @@ namespace SitecoreSpark.SPRK.Tests
             logger.CloseLog(testKey);
 
             Assert.IsFalse(Directory.Exists(this._logFolderPath));
+
+            this.CleanUpTestLogFiles();
         }
 
         [TestMethod]
